+++ examples/aurora-postgresql/main.tf
@@ -21,9 +21,9 @@ module "rds_cluster" {
   engine_mode                     = "provisioned"
   instance_class                  = "db.r5.large"
   subnet_ids                      = data.aws_subnets.database.ids
-  cluster_identifier              = local.cluster_name
+  cluster_identifier              = var.cluster_identifier
   master_username                 = random_string.master_username.result
-  final_snapshot_identifier       = "${local.cluster_name}-snapshot"
+  final_snapshot_identifier       = "${var.cluster_identifier}-snapshot"
   storage_encrypted               = true
   kms_key_id                      = data.aws_kms_key.supporting.arn
   vpc_id                          = data.aws_vpc.supporting.id
@@ -62,16 +62,16 @@ module "rds_cluster" {
 }
 
 resource "aws_backup_vault" "this" {
-  name          = "${local.cluster_name}-backup-vault"
+  name          = "${var.cluster_identifier}-backup-vault"
   force_destroy = true
   kms_key_arn   = data.aws_kms_key.supporting.arn
   tags          = local.tags
 }
 
 resource "aws_backup_plan" "this" {
-  name = "${local.cluster_name}-backup-plan"
+  name = "${var.cluster_identifier}-backup-plan"
   rule {
-    rule_name         = "${local.cluster_name}-backup-rule"
+    rule_name         = "${var.cluster_identifier}-backup-rule"
     target_vault_name = aws_backup_vault.this.name
     schedule          = "cron(0 12 * * ? *)"
 
@@ -83,7 +83,7 @@ resource "aws_backup_plan" "this" {
 
 resource "aws_backup_selection" "this" {
   iam_role_arn = aws_iam_role.backup.arn
-  name         = "${local.cluster_name}-backup-selection"
+  name         = "${var.cluster_identifier}-backup-selection"
   plan_id      = aws_backup_plan.this.id
 
   resources = [
@@ -92,7 +92,7 @@ resource "aws_backup_selection" "this" {
 }
 
 resource "aws_iam_role" "backup" {
-  name               = "${local.cluster_name}-backup-selection-role"
+  name               = "${var.cluster_identifier}-backup-selection-role"
   assume_role_policy = data.aws_iam_policy_document.backup.json
 }