+++ locals.tf
@@ -1,0 +1,3 @@
+locals {
+  vpc_security_group_ids = length(var.vpc_security_group_ids) == 0 && var.create_security_group == false ? null : var.vpc_security_group_ids
+}