+++ variables.tf
@@ -228,11 +228,12 @@ variable "tags" {
 
 #S3 import
 
-variable "s3_import" {
+/*variable "s3_import" {
   description = " Requires that the S3 bucket be in the same region as the RDS cluster you're trying to create."
   type        = map(string)
   default     = null
 }
+*/
 
 # Restore to point in time
 variable "restore_to_point_in_time" {
@@ -393,12 +394,8 @@ variable "family" {
 
 variable "cluster_parameters" {
   description = "(Optional) A list of DB parameters to apply. Note that parameters may differ from a family to an other."
-  type = list(object({
-    name         = string
-    value        = string
-    apply_method = string
-  }))
-  default = []
+  type        = list(any)
+  default     = []
 }
 
 # Enhanced Monitoring
@@ -493,4 +490,10 @@ variable "scale_out_cooldown" {
   description = "(Optional) The amount of time, in seconds, after a scale out activity completes before another scale out activity can start."
   type        = number
   default     = 300
+}
+
+variable "vpc_security_group_ids" {
+  description = "List of VPC security groups to associate"
+  type        = list(string)
+  default     = []
 }